fghf

+++ 03_Dominio/entMenu.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -14,6 +15,6 @@ namespace _03_Dominio
         public entProducto Entrada { get; set; }
         public entProducto Segundo { get; set; }
         public entProducto Postre { get; set; }
-        public int Precio { get; set; }
+        public decimal Precio { get; set; }
     }
 }
